Added bullet points.
Some sections needed legitimate bullet points. These have been added
for future filling in.  A future work section was added. A series of
silly comments involving pizza were added to act as a placeholder for
an actual conclusion -- ideally, our conclusion will be somewhat more
relevant to our work than that of the silly comments, but not yet.