Revised some of the paper.
The introduction has been touched up. The solutions section has been
reworked in light of some of the work we've done so far. A few
bullet points were added or removed. The first couple of paragraphs
of related work has been cleaned up a bit but we still need to address
the remaining related work paragraphs.