Removed sniffer, moved functionality.
The sniffer is no longer needed. A single function in the
controller should be able to replace its ability to convert
packets into scapy-friendly packets.

The controller has a TODO section marked that needs to be
completed ASAP. We need the POX-style function that deals with
packets that arrive on its interface. This function's code has
been commented out but is mostly complete. The remaining step
is to write cleanup code for when the debugger needs to exit.